Instead of doing it the way I proposed in the last commit message, I decided to keep the same generic parent of Core but rename it so as not to be confusing.
